requirements.txt
streamlit>=1.28.0
opencv-python-headless>=4.8.1  # headless version for server deployment
scenedetect>=0.6.2
ffmpeg-python>=0.2.0
requests>=2.31.0
pydub>=0.25.1
pillow>=10.1.0
numpy>=1.26.0
pandas>=2.1.3
google-generativeai>=0.3.1
fastapi>=0.104.1
uvicorn[standard]>=0.24.0  # [standard] includes extras for production
python-multipart>=0.0.6
aiofiles>=23.2.1
python-dotenv>=1.0.0  # for environment variables
